server/pyproject.toml
[tool.poetry]
name = "easytuner-server"
version = "0.1.0"
description = "EasyTuner ECU Map Recognition Platform - Server API"
authors = ["EasyTuner Team <engineering@easytuner.com>"]
readme = "README.md"
license = "MIT"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.11"
# FastAPI and ASGI server
fastapi = {extras = ["all"], version = "^0.104.1"}
uvicorn = {extras = ["standard"], version = "^0.24.0"}
# Database
sqlalchemy = {extras = ["asyncio"], version = "^2.0.23"}
alembic = "^1.12.1"
psycopg2-binary = "^2.9.9"
asyncpg = "^0.29.0"
# Async task processing
celery = {extras = ["redis"], version = "^5.3.4"}
redis = "^5.0.1"
# Authentication & Security
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
# File handling
python-multipart = "^0.0.6"
# Data validation
pydantic = {extras = ["email"], version = "^2.5.0"}
pydantic-settings = "^2.1.0"
# Object storage
minio = "^7.2.0"
# Configuration
python-dotenv = "^1.0.0"
# Scientific computing for detection pipeline
numpy = "^1.26.2"
scipy = "^1.11.4"
# Utilities
python-dateutil = "^2.8.2"

[tool.poetry.group.dev.dependencies]
# Testing
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
pytest-cov = "^4.1.0"
pytest-mock = "^3.12.0"
httpx = "^0.25.2"
# Code quality
black = "^23.12.1"
ruff = "^0.1.9"
mypy = "^1.8.0"
# Type stubs
types-redis = "^4.6.0.11"
types-passlib = "^1.7.7.20240106"
types-python-dateutil = "^2.8.19.14"
types-python-jose = "^3.3.4.8"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
  | migrations
)/
'''

[tool.ruff]
line-length = 100
target-version = "py311"
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "N",  # pep8-naming
    "UP", # pyupgrade
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "SIM", # flake8-simplify
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]
exclude = [
    ".git",
    ".mypy_cache",
    ".ruff_cache",
    ".venv",
    "venv",
    "__pycache__",
    "migrations",
]

[tool.ruff.isort]
known-first-party = ["app"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_any_generics = true
disallow_subclassing_any = true
disallow_untyped_calls = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true
# Allow some flexibility for third-party libraries
[[tool.mypy.overrides]]
module = [
    "celery.*",
    "minio.*",
    "prometheus_client.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --strict-markers --strict-config --cov=app --cov-report=term-missing --cov-report=html"
testpaths = ["tests"]
pythonpath = ["."]
asyncio_mode = "auto"

[tool.coverage.run]
source = ["app"]
omit = [
    "*/tests/*",
    "*/migrations/*",
    "*/__pycache__/*",
    "*/venv/*",
]

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]

